#27 Initial framework
Changes:
RequestParser: Changed it to be a generic handler for requests.
RequestResponseHandler: Generic framework for different apis we suppot.
WorkloadPredictionHandler: Structure of interface to prediction algos.

Workflow:
Different apis that we support have different handlers identified
by their apiName and inheriting RequestResponseHandler.
These have to be registered with RequestParser.
When the parser parses a request, it hands it off to a handler
based on the apiName.

Testing:
This is untested right now