#27 Initial framework
Changes:
Added the rest of the framework.
Added API handler registration and processing code in main.py
Added WorkloadPredictionHandler to serve the 'predictWorkload' apis.
This is pretty much the api that we will need.
Added a dummy test file called "exchange.csv" in data folder for testing of arima.

The architecture is as follows:
We have api handlers for each api.
API handlers will register themselves with the RequestParser.
On each incoming request, RequestParser will extract the 'apiName' from the
request and call the corresponding handler.

Each handler will have a process() and encode() method.
process() will have the functionality of that Handler (prediction in our case).
encode() will construct the JSon response encoding the results that we've
generated in process().

Testing:
Tested the framework on dummy data ("not workloads but exchange rates")
However the rest of the api architecture is in place now.

TODO:
Get data from prometheus and test and improve ARIMA