Increase prometheus tsdb retention
Increase prometheus tsdb retention to 40days.
Achieve this change by creating a init script to specifying the required command line
option.

Flag to increase retention:
--system.tsdb.retention=40d

This flag needs to be command line option
and not a config in the configuration file.

Verification:
Verified the the option is showing up on
the web UI of prometheus
http://127.0.0.1:9090/flags