#54 Added current time
Changes:
1. Added current time instead of the constant time
2. Start time for prediction is 6 hours before the current time.
3. Using wma for prediction by default.

TODO:
There's a weird bug where some of the predictors throw an error for
multiple hosts. This is some parsing issue which needs to be resolved.
So changed the algorithm to wma instead which was working fine when tested.

Testing:
Tested on postman with both the nodes on.

Response value:
[[{"workload": {"172.31.15.135:9100": 5.9802740444085, "
172.31.15.164:9100": 1.1545938375350255}}, {"workloadType": "cpu"},
{"unit": "percent"}]]