#72 Added an api to evaluate prediction algorithms
Changes:
1. Using the root mean squared error metric to evaluate the accuracy of prediction algorithms.
2. Added logging. Logging will happen in predictor.log file.

TODO:
There's a weird bug where some of the predictors throw an error for
multiple hosts. This is some parsing issue which needs to be resolved.
So changed the algorithm to wma instead which was working fine when tested.

Testing:
Tested on postman with both the nodes on.

Request:
{
	apiName: 'getAccuracy'
}

Response value:
[[{"accuracy": {"172.31.15.135:9100": 0.29257265235818103, "172.31.15.164:9100": 0.322202659696561}},
{"metric": "rmse"}]]