Finished the error handling issue #72 by implementing a backoff declarator for the scheduling function
This function stores a persistent scheduling state of with in a time period of all pods that have been
scheduled successfully, this is to prevent wasted overhead of rescheduling a pod that has already been bound to a node.

If a multiple scheduling attempt does occur that results in a conflict api error, the delay to reprovision a pod is
increased.

The following test was conducted:

able to provision multiple pods on multiple nodes by querying the predicator api.

We discovered that using a 15 minute window provided the best results.

Lastly, we discovered that a pod that is pending and has failed to be provisioned will remain
in a pending state.  This is because a pod in a pending state never triggers another event for the
watch module in the python kubernetes library.

To remove a stuck pod, delete and rescale the deployment.