#78 Integrated async threading module for precomputation
Changes:
1. Converted AsyncLoadBalance into a singleton class.
2. getPrediction() now picks the best module from
AsyncLoadBalance.getBestModel()
3. Installing cron inside the docker container.
4. Added an api endpoint /precompute to call the precompute model,
so that the best prediction model is populated periodically
based on the RMSE values.

To configure the precompute module periodically, log in to the
container and install a crontab:
crontab -e
*/5 * * * * cd /usr/bin/src/ && ./precompute_scheduler.sh
This will configure precompute module to run every 5 minutes.

Testing:
Verified the changes on prometheus and k8master.

TODO:
Logs have to be enriched.